+++ src/types/notification.ts
@@ -1,0 +1,5 @@
+export interface Notification {
+  id: number;
+  message: string;
+  type: "info" | "success" | "warning" | "error";
+}
